redesign allocator to reduce memory wastage
now there's two new pieces of configurations:
1. allocator size hint width: number of bytes the allocator puts in
front of an allocated buffer to indicate its size; this length is then
used to index a homogeneous storage instance upon freeing the buffer.
2. allocator free list pointer width: the width of the pointer data
structure used in homogeneous storage, when it frees a buffer the
allocator writes a pointer to the next free block in it, thus the
allocated size must be wider than this pointer.
allocator performs runtime and compile checks to make sure both
constraints are met.