implemented a bucket allocator
Based on Vadim's suggestion, this allocator only handles relatively small items. Allocation and deallocation is done through a free list, supports (supposedly) fast memory-order iteration using a bit mask.
Memory overhead of the bitmask is hard-coded 1/64 of actual memory, and every access is aligned to the size of a power of 2.
Allocation (not growing) and deallocation are both O(1).
Re-allocation behavior can be controlled flexibly through template parameters.